system: core: usb: Remove redundant code
Commit 934dee764d026 ("system: core: usb: Enable start_usb if
SOFTAP_W_DUN is set") added the correct path to read the value
of SOFTAP_W_DUN, which when enabled updates the value to Y.
Because of this, the last executable line of code tries to write
1 to a sysfs entry which is not used anymore returning an error
exit code to the start_usb script. This leads to failure in starting
usb.service and hence USB enumeration fails at bootup when
SOFTAP_W_DUN is enable. This change removes the redundant code
which is leading to this failure. This change also adds proper
checks when qcom_gadget and SOFTAP_W_DUN is enabled.

Change-Id: I943c4af117354e98b8bda3681ea34c3512c6251e
Signed-off-by: Pratham Pratap <prathampratap@codeaurora.org>