system: core: adb: Allow adb root by default if attach is PCIe
Allow adbd to run as root by default when transport is PCIe.
"adb root" command issued over PCIe gets stuck as PCIe cannot
reenumerate its entire UDE stack when adbd is restarted as root.
So allow it to run it as root by default to avoid the command
getting stuck.

Change-Id: Ib5ad130f874231210df88d198ad5cb459ccd7b38