system: core: usb: Exit with success if device is in mhi mode.
When mhi/pcie mode is enabled, boot_hsusb_comp will be set to
none and usb script will execute with this invalid pid. Due to
this the composition switch statement will return an error value
causing usb.service and its dependencies to fail. Avoid this by
exiting with success when mhi mode is enabled.

Change-Id: Iec395692198d04a2bfbf6acad35e5ce34a6508c0
Signed-off-by: Elson Roy Serrao <eserrao@codeaurora.org>