Start usb, adb services as part of default.target
Currently usb and adb services are starting part of multi-user
target which is default when full filesystem is mounted directly
on root(/). But on minimal ramdisk image multi-user target won't
be present and initrd target is going to be default. Instead of
assuming multi-user target availability start these services in
default target so that adb is available always at boot.

Change-Id: I19a47948d4c116001628db32ec4076cbf9df31c5