usb: target: switch qdss usb to sw path on sdxlemur
There is a HW issue on sdxlemur that causes a diag
instruction stall when there is backpressure on ATB.
If QDSS-USB is in bam-to-bam mode, ATB is put under
backpressure frequently and would cause a modem
crash due to the above. To solve this, switch qdss
usb to software path on sdxlemur.

Change-Id: I7fbf943307c93bb741c3eda671ee8916a0c07049
Signed-off-by: Rohith Kollalsi <quic_rkollals@quicinc.com>