logd: Do not set uid gid to AID_SYSTEM
- Remove UID and GID setting to AID_SYSTEM in reinit thread

POSIX requires that all threads in a process share the same
credentials. When the reinit thread get to execute first it
sets UID and GID to SYSTEM, all of the other threads in the
process also change their credentials due to this. It cause
all the system calls that need process to run as root initially
and as logd later to fail and logd exits unexpectedly.

Change-Id: I63318d2d95aea0aa1d0ba00779774b13e0c0165b