core: usb: Bind UDC to ConfigFS when UDC gets added
In LE enviroment, when we bootup in Host mode, UDC will not be
present, hence start_usb script wont be able to bind UDC to
configfs. Now when we switch to peripheral mode, UDC will be
available, but there is no one to bind it to configfs.

When the DuT is in host mode, and peripheral gets disconnected,
we send a "remove" uevent for each device. This is monitored by
usbd and extracts the udc name from the uevent. And this string
is written on configfs UDC, which will bind UDC with configfs.

Change-Id: I18bd5aff4353be9c206ad2fda4bf409a6aab7bae