core: usb: Optimize CPU usage of USBD
USBD monitors all the kernel uevents and checks for USB events
from them. Currently we process every line of all the kernel
events, which is very inefficient as it consumes lots of CPU
cycles. Optimize this by checking only the first line of uevent
which contains the data we are interested in.

Change-Id: I134ded88241be769cc6de1c59e1d2096e1818741
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>