core: start_usb: Move acm.0 & enable console for acm.0
Rearrange the creation of interfaces in such a way that acm.0 is
created after gser.0, so that ttyGS0 will be assigned to gser &
ttyGS1 will assigned to acm. This is done to avoid discrepancies
on g_serial interface, like some userspace applications might be
dependent on ttyGS0.

Also enable console for acm.0 interface so that the kernel logs
will be printed on the the new ttyGS1 console.

Change-Id: I3743e1787e8ee8b5c28ea876c6ebd4df1111f4c2
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>