start_usb: Set usb composition when switching from PCIE to USB
Currently we set the composition to none when the target is in
MHI mode. And we don't set composition if we are switching from
MHI to USB, and this was causing USB enumeration to fail.

We removed overwriting boot_hsusb_comp in start_pcie from commit
d554466e8c4fb (adb: start_pcie: Avoid overriding composition
from start_pcie) because setting boot_hsusb_comp from start_pcie
was overwriting the persist USB composition.

And now if the composition is 'none' we set boot_hsusb_comp to
default composition in start_usb based on target name.

Change-Id: I444b20412ab48293bbdde15ae604896b107497b7