system: core: usb: Restart adbd if failed due to UDC sysfs
Configfs based USB composition fails when switching from
composition done by the qti_usb gadget driver. The reason
for the failure is that the adbd is not start when booting
with early USB configuration. The UDC entry in sysfs
(/sys/kernel/config/usb_gadget/) is not appear if boot with
an early USB. So the ExecStartPre of the adbd service fails
due to the unavailability of the UDC sysfs entry.

These changes fix this failure when switching to configfs
based USB composition.

Change-Id: I447863de1a99e82e266777bc4eadb2f47d1549fe
Signed-off-by: Selvarasu Ganesan <quic_selvaras@quicinc.com>